pyproject.toml
[project]
name = "tryna-master-rag"
version = "0.1.0"
description = "markdowns, advanced chunking strategies, and more"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
]
